Load viaduct.js as a separate file
Loading viaduct.js as a separate file makes it much easier to debug it
with Firebug, or the like.

Doing this, however, means fetching the gateone endpoint URL separately
from the server, as well.  It would also be possible to load that
variable from the HTML template, but I have the impression that fetching
it asynchronously will make things more flexible for the future.