Wait to connect until environment is provisioned
This makes the xblock wait (for 2 minutes, currently) until the environment is
SSH-able before telling gateone to connect to it.  This fixes hastexo/viaduct-xblock#10.

It does so with the cooperation of the Heat template, which sets the training
user's shell to /bin/false while Ansible is running.  As a very last command,
cloud-init then restores it to /bin/bash, allowing the training user to SSH in.

In order to test the SSH connection, a way had to be found to give the xblock
access to the same passphraseless private key gateone would use.  Instead of
creating a hard-to-maintain external depency, it was decided to implement the
"one key per user" feature already.  This also requires the cooperation of the
Heat template, which outputs a private key that can be used by both gateone and
the xblock checking procedure.  This fixes hastexo/viaduct-xblock#7.