(Partly) Fix field scopes
Fix the field scopes as per the documentation.  The stack configuration
is set course-wide, so most of the settings are Scope.content.

What's missing is to fix the user_state scope so that related variables
are set per course, per user.  There doesn't seem to be a predefined
field for it, so we need to figure out how to define it.