Try to resume stack on get_user_stack_status()
If the student takes long enough to get to the terminal in a section
with a lab, the stack might get suspended in the meantime.  Once he does
get there, the stack needs to be resumed.  To do so, try resuming it
again when the browser requests the stack status.