Improved idle detection and UI
Introduced here is idle detection based on keyboard input in the browser
terminal.  By default, if the user doesn't type anything in 10 minutes,
their stack will be suspended.  When that happens and their browser
window is showing the terminal unit, a modal dialog will be displayed:
resuming the stack will require clicking a button.

If the student navigates away from the lab unit but remains in the same
section, the idle and keepalive timers still run.  The result is that
their stack will be suspended only after the idle timer expires (as
opposed to when the shorter keepalive timer does).  This is intentional,
as it is better UX for the student to be have more time to return to the
lab unit and not need to have the stack be resumed.

The dialog itself is a new feature introduced here, and is also used in
interaction with the user elsewhere, such as during stack launch, or
running assessment tests.  An unintrusive, pure-CSS loading animation is
also included for situations where the student needs to wait.

This changeset also fixes a bug where the "Check progress" button would
disappear when navigating away and back to the lab unit.

Finally, there was some refactoring of the Javascript code:

* a new plugins.js file contains code used by, but not a direct part of,
  the XBlock (such as the dialog plugin).

* a few readability improvements in main.js