Do away with infinite loops
There were a couple of edge cases where tasks would loop indefinitely.
For example, if a stack was deleted while the task waited for it to be
created or resumed.

This changeset treats this case specifically, but also gets rid of any
infinite waiting by introducing timeouts where there weren't any before.