Use correct scope to store stack state
The student's stack is the same across uses of the hastexo XBlock in a
course.  As such, certain state variables should be using the
appropriate scope for this.

The closest thing to what's needed is the "preferences" scope.  However,
the documentation states that it stores information across courses as
well.  It remains to be seen if this is actually the case.