Only launch stack when on lab unit
In order to simplify the launching logic and to make it easier to test,
don't launch stacks when rendering the student view, as this is
something that happens when the sequential is rendered, not when the lab
unit is displayed.

In other words, from now on, the student stack will only be launched
when the student views the lab unit.