Distributed keys
Up until now, it was necessary for RabbitMQ, the Celery workers, and
GateOne to run on the same app server: it was the only way to guarantee
that the SSH key saved by the Celery worker that launched a stack
would be accessible by the GateOne SSH connection script.

With this changeset, this is no longer necessary.  By saving SSH keys to
Swift, workers and GateOne are fully independent.  The launching worker
will upload the generated key, and the modified SSH connection script
(introduced here) will download it from wherever it runs.  (Note that
this behavior is optional.)

However, due to the necessity of having a system-wide configuration to
access the OpenStack cloud, it is no longer possible to set
OpenStack authentication parameters per XBlock definition.  These must
be set globally on lms.env.json.

As a bonus, all timeouts are now configurable via XBLOCK_SETTINGS.

Note: for this to work, the new connection script must be set in the
GateOne configuration as well.