Optimize time-to-terminal
This improves the get_stack_status algorithm so that the time from
entering the unit and getting a connected terminal is minimized,
in particular in certain edge cases:

Case 1
------

When the browser requests a stack's status and the server finds it to be
COMPLETE, we still need to be reasonably sure that the stack hasn't
suspended since the previous request, before giving the browser the
go-ahead to connect to it.  There are several possible reasons this
could happen: the user's computer entered suspension, the user closed
the tab with the XBlock, there was a temporary hiccup in the
connectivity between the browser and the server, etc.

Previously, we worked around this by always running a second
launch/resume task after the first one returned, this time
synchronously, making sure no suspend tasks were scheduled.  While this
worked, it was always perfectly possible that this would take an
unreasonable amount of time, resulting in Ajax timeouts and other
problems - problems that are supposed to be avoided by asynchronous
Celery jobs in the first place.  Notably, one of the possible problems
would be a failure to schedule a suspend task for the launched task.

Instead of doing the above, when a launch task is found to be
successful during a status check, simply check if it's reasonable to
assume (even if not necessarily true) that the stack was suspended.  Do
so by considering how much time has elapsed since the last suspend task
was launched (which coincides with the last time the status was
checked).

If it's likely the stack was suspended, which implies a large hiatus
since the last check, simply start from scratch: run a new asynchronous
launch/resume task (including resetting the global timeout), report the
launch as pending, and proceed as usual.

If, on the other hand, not enough time has elapsed since the last status
check, it is safe to a) assume the stack is still alive and kicking, and
b) to give the go-ahead to the browser to SSH into it.

Case 2
------

This changeset also solves a related edge case, where the user would
open a new XBlock unit, a long time after closing a previous one while
it was still being launched.  If this happened, the user would get a
timeout notification no matter what the original launch task status
ended up being.  Now, opening a new lab tab will always force a new
launch task to be issued, with the previous one forgotten.  Unless, the
next case is encountered, however:

Case 3
------

If the user leaves a lab (by closing the tab, or just going to another
unit) and returns to it in under the time it takes the suspend task to
fire, the XBlock just assumes the stack is still ready, doesn't query
Heat, and just allows Gateone to connect promptly.