Store stack state per course run
Emulate a "per course run, per block, per user" XBlock scope by storing
stack-specific state in a Scope.preferences dictionary (this scope applies "per
block, per user").  This is effectively "per run" because stack names are
built to be unique per run.

This solves a particularly problematic issue: previously, if a learner went
quickly (i.e., took less than the suspend timeout) from one lab to another in
the same course, the stack would be suspended mid-use because the first lab's
dead man's switch would not be revoked.

As an added bonus, if the student does go quickly from one lab to another, the
time-to-terminal is as rapid as leaving a lab and going back to it under the
suspension timeout.  (i.e., there is no need to query Heat).

Note: This changeset paves the way for multiple stacks per user, per
run.  If this feature is desired, all that is need is to change the
stack naming algorithm to produce different names for different
templates.