Revisit stack launch logic
Previously, if a user refreshed the page or opened a second lab tab
while a stack was being launched, a second concurrent launch stack would
be fired.  This had the detrimental effects of a) maximizing the
probability of race conditions, and b) delaying the original launch.
This rewrites the launch logic so it can't happen.

In other words, the user can open as many tabs as they want (including
to different instances of the XBlock in the same course), or refresh as
often as they can, and it will have no effect on the ongoing launch.
Until an error or a timeout is encountered, that is; at this point, a
refresh will start a new launch task, as usual.

This changeset also strives to make the logic clearer by hinging
primarily on the saved stack status string - as opposed to *both* status
string and the launch_id.  The latter is now ignored when deciding
whether to launch a new task).

Included is a reorganization of method placement, for readability.