Multiple cloud providers
Reimplement the ability to specify a separate cloud provider for each
XBlock instance.  The author can't set specific credentials, but is
allowed to pick a provider name from a list of providers configured at
the platform level.

If no provider is set, the "default" one will be used.  If there is no
configured "default" provider in the platform, the first one will be
used.  If there are no configured providers, the XBlock looks for a
previous version of the configuration, where only a single set of
credentials was supported.

SSH keys are stored at and retrieved from the same provider where the
learner stack was launched.