Don't withhold paste keys indefinitely
It's not possible to predict when a `paste` event will happen.  Certain
browsers may not support the paste shortcuts predicted here, with the
result that the final key stroke will never be sent remotely.

Instead, insert a small delay on the final keystroke, a delay which
needs to be only as long as it takes for the browser to register the
`paste` event.