Marginally improve the client's handling of Guacamole server errors
When the Guacamole web socket returns an error, only display the full
error message if it is *not* the "Aborted. See logs." message (which
is entirely unhelpful to learners, who presumably do not have access
to those logs.)

If, however, it *is* "Aborted. See logs." (or to be precise, any error
message whose lowercase representation starts with "aborted"), then
inform the learner that they might be able to get themselves out of
the problem by closing the browser, and waiting for a few minutes
before retrying. Under the assumption that the XBlock has been
configured to use a suspend timeout of two minutes or thereabouts, the
retry will then trigger a full resume cycle.