Separate suspender program
Store user stack data in a separate table in the database, so it can be
properly locked to avoid clobbering, and also so a separate process can
suspend stacks independently of user requests.

The suspender process is implemented as a custom django command, and can
be invoked with, for example:

```
./manage.py lms --settings=openstack suspender
```

Note that if the suspender isn't running, stacks won't be suspended at
all.