Fix the reaper's "MySQL has gone away" error
When set to run with an interval of 6 hours, after a few runs the reaper
would start failing continuously with a "MySQL has gone away" error.
This is due to the persistent connection dropping out between runs, and
Django not reconnecting by default.

To avoid this, close the connection after every run.