Update provider in real time
The launch task can take a long time.  If we wait until it finishes (and
until the user comes back to check!), the capacity calculation algorithm
is hopelessly broken in busy environments - precisely the situation
where it should be most useful.

Therefore, the stack provider must be updated in the database as soon as
a provider is attempted.