Deprecate custom XML parsing
The XBlock's built-in XML parser supports an "option" namespace, which,
via YAML parsing, obviates the need for the current custom `parse_xml`
method.  With it, it is not only less cumbersome to write the XML, but
it is now exported cleanly (as opposed to the current JSON-in-a-string
mess for the custom fields).

This deprecates the current custom parsing engine, but keeps it for
backward compatibility.