Multi-cloud support
Refactor the codebase so that provider-specific code is separate from
stack launching logic, in such a way that tasks and jobs need not know
anything about the underlying cloud.

Add a "type" server-side configuration for providers (currently, just
"openstack").  For backwards-compatibility, defaults to "openstack" if
not set.

Also, let course authors specify templates per provider.  For backwards
compatibility, if per-provider templates are not given, use the global
`stack_template_path`.

Rename `utils.py` to `common.py`, as it more accurately reflects what
the file is for.