Improve confirmation message on reset
Based on feedback from multiple users, it appears that we're not
being clear enough about the fact that a lab reset is a lab reset,
i.e. that it returns the learner to square one and not just the
prior lab.

Change the warning message so that it's unequivocal that the learner
is about to reset their lab to a pristine state, and hammer the
point home that they will have to redo everything they've previously
done.