Add Python 3.6 to test matrix
With the previous changes to the test requirements (notably, using
tracking xblock-sdk master), we can now enable Python 3 testing.

However, we do suffer from an issue where we're using binary input
with JSON, something that wasn't allowed in Python 3.5 but is as of 3.6:

https://docs.python.org/3/whatsnew/3.6.html#json

Thus, until we decide whether we want to make our code play nicely
with 3.5, or else drop 3.5 support altogether, we'll make the build
chain break on Python 3.6 errors, but ignore errors on 3.5.