Avoid known stack suspension failure states
1. When servers in a stack are rebooting, such as immediately after
resuming, Heat will refuse to suspend the entire stack.  This leads to
the stack being put in a failed state which can only be resolved
manually.  To avoid this, don't try to suspend the stack at all when any
servers are in such a state, and instead, retry later.

2. Most cases of suspension failure can be fixed by simply retrying.  To
do this, include `SUSPEND_FAILED` stacks in the list of stacks to try
suspending again.