Reaper zombie destroyer
It so happens that the Heat API, when under yet-undetermined but clearly
adverse conditions (usually after long-running, failed launch attempts),
will report a 404 for a stack that is actually still there.  This causes
the XBlock state machine to remember that stack as deleted, with the
undesireable effect that if the owning user never tries to launch that
stack again, it will never be considered for reaping.

To avoid having such useless stacks lying around accruing running
charges, this implements a catch-all phase to the reaper run.  This
phase will list all stacks on each configured provider and compare each
stack to its corresponding state in the database.  If the stack is
deemed to be a zombie (i.e., it is known to be dead, but clearly isn't),
it is immediately destroyed.