Fix XML parsing backward compatibility
From the introduction of per-provider stack templates on v3.0.0, the
already-deprecated `parse_xml` method started requiring a template
definition for each provider.  This is inconsistent with the new
option-handling code, which handles this at run-time, in a
backward-compatible manner, by falling back to the global
`stack_template_path`.

This removes the template requirement at parse time, so that providers
defined the old way continue to work as expected.