tasks.py: continue on specific EnvironmentErrors during SSH connection
When attempting to connect with Paramiko, we might not yet have full
connectivity to SSH; in particular, a virtual router may not be fully
operational yet. Under such circumstances, we might encounter a
socket.error, such as error 104 "Connection reset by peer".

Previously, such a socket.error would cause a stack launch (or stack
resumption) to needlessly fail.

Because of all the changes to OSError and socket.error between 2.7 and
3.4+, unit testing with it is a flaky affair.  However, given that:

* we still need to support Python 2.7

* from Python 2.6 until Python 3.3 socket.error is a subclass of
IOError, which in itself subclasses EnvironmentError, which implements
errno,

* from Python 3.3 onwards socket.error is an alias for OSError, of which
EnvironmentError is an alias, which also implements errno,

* we only care about some errnos anyway,

we handle socket.error via its common parent denominator of
EnvironmentError.

In addition, add more logging to SSH connections.

Co-Developed-by: Adolfo R. Brandes <adolfo.brandes@citynetwork.eu>