Use empty JSONFields instead of "null"
This is another attempt to get JSONFields to behave correctly (and
expectedly) when they're unset.  Previously, empty `hook_events` and
`providers` were incorrectly set to "null" - including quotation marks.
This would lead to instances being interpreted by the Django ORM as
strings, instead of the expected Python `NoneType`.

To fix this, set the fields to empty dicts or lists as appropriate.
During the migration, also update the rows accordingly.