Refactor stack state variables
To minimize confusion, refactor stack state variables into a consistent
naming scheme.

In doing so, expand valid states to include all foreseeable stack
states, and also include the empty string in the admin form state
dropdown.  From this commit on, if the status dropdown is empty in the
admin page, the stack is in an unforeseen state and should be
investigated.