Fix error handling when waiting for state change
When waiting for a stack to change state prior to creating or resuming
it, take into account which of the two is happening.  Otherwise, a
resume failure here could result in a stack's provider being wrongly
cleared.

Also remove an unreachable line in DeleteStackTask, discovered due to
the increased unit test coverage added here.