Include SUSPEND_COMPLETE in OCCUPANCY_STATES
When assessing whether a provider is at capacity or not, we would not
include suspended stacks in the count. This meant that for stacks that
use automatic suspend (where most stacks are suspended at any given
time), the only trigger for using the non-default provider would have
been for a stack launch to run into an error.

Instead, include SUSPEND_COMPLETE in OCCUPANCY_STATES, making them
count toward provider utilization.

Also, modify the unit test simulating that a provider is full, so that
it deals with suspended rather than freshly created stacks.