Close database connections before long-running operations in Celery tasks
Launching, suspending, resuming, and deleting stacks are potentially
very time-consuming operations, so much so that they might span a
database restart or failover, with a database operation required
either side. For example:

1. We set a stack state to LAUNCH_PENDING.
2. We fire off an API command to create new a stack.
3. The database server restarts.
4. Launch completes.
5. We set the stack state to CREATE_COMPLETE.

Or:

1. We set a stack stack to DELETE_PENDING.
2. We fire off an API command to delete the stack.
3. A load balancer sitting in front of a database cluster fails over
   to another backend.
4. Deletion completes.
5. We set the stack to DELETE_COMPLETE.

In either case, item 2 takes on the order of several minutes. Item 3
tears down a database connection that would otherwise persist for the
task's lifetime, so while item 1 succeeds, item 5 fails with an error
indicating that the server has disappeared mid-session (for MySQL,
that would be error 2006 in the first scenario, and 2013 in the
second).

Thus, instead close the database connection immediately before item
2. For item 5, Django will automatically reconnect.