Disable page auto-scroll when terminal gets focused
In order to interact with the Guacamole terminal in the browser window,
we need to focus the element, when we mouse over it.
Focusing the terminal however, in some browsers, will automatically
scroll the page to fully display it. This behaviour can be distracting
for the students and make it difficult to read text surrounding the terminal.

This commit disables the auto-scrolling behaviour for Chrome
(and presumably all Chromium-based browsers).
The behaviour in Firefox and Safari are not affected by this change.
Firefox is not auto-scrolling when terminal is focused by default.
Safari however, is auto-scrolling and this commit does not disable that.

Our goal is to get Chrome and Firefox behave in the same manner.