Allow overriding 'delete_age' by defining an XBlock attribute
This commit adds two new fields to the Stack model: 'delete_age' and 'delete_by'.

When a lab gets suspended, a 'delete_by' timestamp will be added to it, which will be
calculated by adding the 'delete_age' amount of time to the moment of suspension.
The reaper will use the 'delete_by' timestamps to find and delete old lab stacks. The
'delete_by' value can additionally be overridden in the admin panel.

So far, 'delete_age' has been configurable only via XBLOCK_SETTINGS.
This change makes it also configurable via XBlock attributes. If a value
is set to the attribute, it will override the global setting. The XBlock attribute
supports values in seconds.
Currently, the 'delete_age' global setting still only supports days and is converted
to seconds internally. In future releases the setting will begin to support suffixes:
`d`, `h`, `m`, `s`.