Fix hook_events import
hook_events is a dictionary that contains boolean values. In course
export those values are converted to strings, but were not properly
parsed back to boolean values in course import.

Use distutils.util.strtobool to make sure that the boolean values in
hook_events are properly parsed in course import.