Add an option to display a terminal in read-only mode
This commit adds an optional XBlock attribute "read_only",
that allows to display a lab terminal in read-only mode.
If set to True, a stack will be created or resumed as usual,
the student can see the terminal but is unable to interact with it.
The value defaults to False.