Support ping checks for IPv6 stacks
On Ubuntu Xenial (the reference platform for Open edX Juniper and
earlier), the ping utility does not support IPv6. For stacks whose
public_ip is an IPv6 address, we have to use ping6 instead.

Thus, parse the public_ip output to determine if it's an IPv6 address
and use ping6 in that case; otherwise use ping.

For regression testing purposes, add IPv6 testing for all unit tests
in test_tasks.py, so as to make sure that they all work with both IPv4
and IPv6.