Fix exporting provider attributes
Provider attributes like `template` and `environment` will not be
added to the provider node during course export in case of None or
"None" values.
However, not defining those attributes at all will throw an error
during export.
Add an additional check to make sure export succeeds when these attributes
are not defined for a provider.