Retry remote_exec on SSHException
When we run remote_exec(), such as when executing a suspend or
resume hook, we might run into a transient error, such as the remote
host (or an intermediary router) suddenly tearing down the connection.

In that case, we don't want to give up immediately, but we also
don't want to retry indefinitely. Thus, use the retry decorator to
attempt the command up to three times, waiting 10s in between.

Co-authored-by: Maari Tamm <maari.tamm@citynetwork.eu>