Retry on Paramiko socket timeout
When we don't get an SSH connection from Paramiko immediately, we
don't want to wait forever until we hit the task timeout. Instead,
we define a shorter timeout, ssh_connect_timeout, defaulting to
10 seconds, and use that to set a socket timeout option with the
Paramiko connect() call.

If we fail to establish a connection within that time, we hit the
socket.timeout error, which we then catch in order to retry the
connection. Also, bump the log level associated with such messages,
to info.