Fix test requirements for CVE-2021-3281
Django 2.x releases prior to 2.2.18 have an insecure
django.utils.archive.extract method. We only install Django in
testing, and the problem appears to affect only "django-admin
startproject" and "manage.py startapp" which are obviously never used
in day-to-day operations, but we should still fix the dependency.

Bump the django dependency in test.txt to ">=2.2.18,<3.0".

Reference:
https://github.com/advisories/GHSA-fvgf-6h6h-3322