Exclude abstract methods from coverage
In abstract base classes, we have methods that do nothing but raise
NotImplementedError, because they are meant to be overridden by
subclasses. Those methods obviously never get any unit test coverage
(nor are they meant to), so use the approach outlined in the Coverage
documentation to exclude such lines from the report.

Reference:
https://coverage.readthedocs.io/en/coverage-5.5/excluding.html