Add default values for Stack.key and Stack.password
In hastexo.models.Stack, we have a key and a password field that
only set blank=True, but neither null=True nor a default.

In Open edX Juniper, which built on MySQL 5.6, this would result
in an "implicit default" of the empty string being applied.[1]

However, as of MySQL 5.7.5[2] and MariaDB 10.2.4[3], the default SQL
mode is STRICT_TRANS_TABLES, which disables implicit defaults
for the InnoDB storage engine. Thus, calling save() on a Stack
object would result in IntegrityError if either the key or the
password field are unset.

Update the model so that it sets a default value of ""
(empty string) for both fields.

References:
[1] https://dev.mysql.com/doc/refman/5.6/en/data-type-defaults.html#data-type-defaults-implicit
[2] https://dev.mysql.com/doc/refman/5.7/en/sql-mode.html#sql-mode-important
[3] https://mariadb.com/kb/en/sql-mode/#strict_trans_tables

(cherry picked from commit d6fd424485b99b2937063564569d9f7aa0a855d4)