Refactor closing ssh connections
It may happen that a target stack is unexpectedly unavailable, either
due to a networking problem or the target VM is killed outside of the
XBlock's own functionality.

When that happens, a task can hit its timeout, fail to create an
ssh connection and leave the `ssh` variable undefined.

This will cause the SSH connection cleanup to break in the finally
clause.

Check if an ssh connection exist before trying to close one to
avoid any unhandled exceptions thrown in this scenario.

(cherry picked from commit d6069df92b7d9bed5a9964645eec8243f28613d9)