Fix dependency updates
When running the OpenStack libraries on the Train release versions,
we ended up having dependency conflicts with `edx-platform`, that
sets `stevedore==1.32.0`.

Add constraints to `dogpile.cache`, which is a requirement for
`openstacksdk` and adds `stevedore>=3.0.0` in version 1.0.0.

Add constraints to `cliff` which is a requirement for python-heatclient
and adds `stevedore>=2.0.1` in version 3.4.0.