For testing, install xblock-sdk from PyPI
For historical reasons (namely, the fact it didn't get any PyPI
releases from 2014 to 2020) we used to install xblock-sdk from GitHub.

As of release 0.1.9 (June 2020), that is no longer necessary.

Since release 0.3 makes use of f-strings that are not supported in
Python 3.5, test against xblock-sdk<0.3 in the py35 testenvs, and
against the latest xblock-sdk release in py38.

Reference:
https://pypi.org/project/xblock-sdk/#history