Sanitize stack names
We build the `stack_name` based on `course_id` and `student_id`.
We also use the same `stack_name` for creating a keypair for the
learner. However, Nova has a list of "unsafe" characters which
will throw an exception, if used in the keypair name.

Make sure that the `stack_name` consists of only safe characters.