Drop wait_for_ping() method to support containerized deployments
When deploying the edx-platform with Tutor, the functionality
for the wait_for_ping() method breaks as `/bin/ping` is not
available in the Tutor Docker containers. Any pure-Python solution
would also require either root privileges or raw socket capabilities,
neither of which is also available.

We really don't need to rely on this logic as we are more interested
in the abilty to create an ssh or rdp connection to the lab.
We can drop the logic for it altogether.