Allow moving on to next provider if one is not configured
Don't fail to run if one from the list of providers is not
configured. Allow moving on to the next and log a warning message
about failing to initialise a provider.
If none of the listed providers are configured however, log an
error message and raise an exception.

(cherry picked from commit 423e03f8abdc3fb4fa8f99c4615413b1ea677610)