Add a check to the `add_user_foreign_key` migration
It may happen that there are some problematic stacks in the database
that would cause the migration to fail when the stacks cannot be
linked to a user account. These stacks are most likely a result of
testing with staff access, overriding some requirements that learner
stacks are subject to.

Add a check to the migration to find out if such stacks exist and
if they do, raise an exception with a message providing the list of
problematic stack names and guidance how to proceed.