Fix the add_user_foreign_key migration
In some circumstances, the add_user_foreign_key migration would fail,
specifically when the stack cannot be linked to an existing user
object. This might occur if the user account has been deleted, or if
the stack was spun up by a Staff user, without enrolling into the
course first.

Fix the migration scripts, so that this issue does not recur.

Co-authored-by: Maari Tamm <maari.tamm@cleura.com>