Add additional custom actions to the admin interface
Add a custom actions to the admin interface to:
- clear the stacklog for selected stack(s)
- set status to SUSPEND_COMPLETE for selected stack(s)

We might need to clear the stacklog of a stack in special cases,
for example when we need to "reset" a lab usage timer for a learner
as the time is accumulated based on stacklog entries.

Setting a status to `SUSPEND_COMPLETE` is a common fix for
stacks that end up in unexpected states.