fix: Restrict twisted to a version below 23.8.0
The channels package depends on daphne, which in turn depends on
twisted>=18.7. Officially, twisted's 23.8.0 release is the last to
support Python 3.7, so it ought to be fine to use, but
"tox -e pipdeptree-requirements" fails on Python 3.8 and 3.9 while it
succeeds on Python 3.10 and 3.11.

So, evidently the twisted 23.8.0 release (which landed on 2023-08-28)
includes some transitive dependency on a package that has dropped
support for Python 3.8 and 3.9.

Thus, restrict twisted to releases prior to 23.8.0.
We can drop this restriction once we decide to drop Python 3.8 and 3.9
support.

Reference:
https://github.com/twisted/twisted/releases/tag/twisted-23.8.0