docs: Add version reference and compatibility matrix to README
It probably does not hurt to include the current version in the
README. To make sure that it gets proper updates upon the release of
a new version, add a reference to the README to the bumpversion
configuration, as well.

In addition, add a version compatibility matrix for older Open edX
releases.