docs: Fix README reference to stack_protocol
The parameter which we use to specify the protocol to connect to
the stack has always been named "stack_protocol" since its inception
in 5d8fa14729a5fb90f1a31e8864542b43af65432a.

For some reason, it ended up in the README as "protocol" in
f82d0d83e06cc272a5b89a19e23fa603d671cee2, and this was never caught or
fixed.

Fix that reference, and also move it from the list of required
parameters to that of optional ones (since it has a default, "ssh").