refactor: Use importlib.metadata to define version, support Python 3.12
We previously made use of the pkg_resources module to determine the
package version (as defined by setuptools-scm) in __about__.py. This
module is included in setuptools, which as of Python 3.12 has been
unbundled from Python, which means it is no longer installed by
default in a freshly initialized venv.

There are multiple ways to address this, one being to include
setuptools in the install_requires list for Python 3.12 and
later. Another, also listed in the Python Packaging User Guide, is to
instead use importlib.metadata.version. Since this approach works for
all Python versions we intend to support (that is, from 3.8 forward),
this is the better solution.

Thus, refactor __init__.py to use importlib, and add Python 3.12 to
the test matrix.

For this to work correctly, we also need to (finally) allow a version
of setuptools-scm>=6 in setup.py.

Reference:
https://packaging.python.org/en/latest/guides/single-sourcing-package-version/