Update test matrix
xblock-sdk v0.9.0 includes a breaking change for XBlock versions
before 2.0 but there is no upper constraint included in the
requirements for this for all XBlock versions. Thus, all our tests
have been failing since the release of xblock-sdk==0.9.0 on
2024-03-21.

Update our test matrix to include XBlock versions 1.7 / 1.8 / 1.9
and add an upper constraint for xblock-sdk.

The Open edX Quince release is using XBlock v1.8, which does pin
the xblock-sdk to v0.7.0 so in this case we only need to add the
constrait to our test runs.