Fix the way we evaluate correctness of the XBlock
Use the score value to define correctness for this XBlock
instead of the progress check result.

When masquerading as a specific learner we are not able to override
a progress check for the learner but as instructors we are able to
fix the score for a learner.

The XBlock should also be evaluated as 'correct' when an instructor
has overriden a score for the learner.